Fix BC7 modes 0, 1, 2, and 3 refinement step when the source data supplies an alpha channel. (Determinism/correctness)
The `bc7_enc_mode01237` subroutine was passing `state->channels` (i.e. number of channels for the whole image) downstream instead of the local variable `channels` (i.e. number of channels for the specific BC7 mode we are attempting). The alpha channel is usually ignored (i.e. left uninitialized), so not ignoring it in this case meant reading potentially-uninitialized stack memory and using it for the error metric.