ASTC: Fix solid color bug in ls_refine_scale
There was already an attempt made to fix a problem with solid-color
blocks that tries to address the case when all weights are equal
but the system has a non-0 determinant. It did not address the
case when all weights are identically 0 which makes us come out
with a determinant that is exactly 0, so sgesv2 actually returns
false.

In that case, we would continue with scale=0, even though all
pixels currently use the same weight of 0, which picks the first
endpoint.

The first endpoint is the one that "scale" gets applied to.
Therefore, we'd then end up making the whole block black, no
matter what was in the pixel data.

Needless to say, this is bad. Scale=1 is a much saner default
here since it tries to keep us as close to the average as we
can manage to hit.

Also add some comments explaining what is going on in this function.
I had to work this out from the code and it seemed worth writing
down.

Signed-off-by: Fabian Giesen <fabian.giesen@epicgames.com>